python/src/auth/requirements.txt
astroid==3.1.0
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.1.2
blinker==1.7.0
boto3==1.28.62
botocore==1.31.62
cffi==1.16.0
click==8.1.7
colorama==0.4.6
comm==0.1.2
cryptography==42.0.5
debugpy==1.6.6
decorator==5.1.1
dill==0.3.8
executing==1.2.0
Flask==3.0.2
Flask-MySQLdb==2.0.0
ipykernel==6.21.2
ipython==8.10.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jupyter_client==8.0.3
jupyter_core==5.2.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mysqlclient==2.2.4
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==4.2.0
prompt-toolkit==3.0.37
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.8.0
pylint==3.1.0
PyMySQL==1.1.0
pyperclip==1.8.2
python-dateutil==2.8.2
pyzmq==25.0.0
s3transfer==0.7.0
six==1.16.0
stack-data==0.6.2
tomlkit==0.12.4
tornado==6.2
traitlets==5.9.0
urllib3==2.0.6
wcwidth==0.2.6
Werkzeug==3.0.1
